+++ openstack/types.go
@@ -20,6 +20,7 @@ import (
 	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/fwaas/rules"
 	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/layer3/floatingips"
 	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/layer3/routers"
+	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/subnetpools"
 	"github.com/gophercloud/gophercloud/openstack/networking/v2/networks"
 	"github.com/gophercloud/gophercloud/openstack/networking/v2/ports"
 	"github.com/gophercloud/gophercloud/openstack/networking/v2/subnets"
@@ -350,4 +351,10 @@ func (opts ZoneCreateOpts) ToZoneCreateMap() (map[string]interface{}, error) {
 	}
 
 	return nil, fmt.Errorf("Expected map but got %T", b[""])
+}
+
+// SubnetPoolCreateOpts represents the attributes used when creating a new subnet pool.
+type SubnetPoolCreateOpts struct {
+	subnetpools.CreateOpts
+	ValueSpecs map[string]string `json:"value_specs,omitempty"`
 }